api-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
anthropic==0.7.7
httpx==0.25.2
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
boto3==1.34.0
Pillow==10.1.0
